feat: Add export functionality and UI elements
Introduces a new feature to export the current list of SystemD services to a text file. This includes adding a new "Export" button to the UI and implementing the necessary IPC handlers in the main process to trigger a save dialog and write the service list to a user-specified file. Also adds rudimentary filter buttons to the UI.